feat: add ssl support for the API
New config options have been added to support HTTPs conenctionto the
API:

```
[http_api]
ssl_enabled = false
ssl_cert_path = "./storage/ssl_certificates/localhost.crt"
ssl_key_path = "./storage/ssl_certificates/localhost.key"
```
